修改消息传递协议,重写proto格式

+++ lib/info.go
@@ -7,7 +7,7 @@
 package lib
 
 type Info struct {
-	Id  int64  `json:"id"`
-	Sid int64  `json:"sid"`
+	Id  int32  `json:"id"`
+	Sid int32  `json:"sid"`
 	Msg string `json:"msg"`
 }
